fix: page <-> index 파일명 수정
