+++ app/kakao/oauth/index.tsx
@@ -1,7 +1,43 @@
-import dynamic from 'next/dynamic';
+'use client';
 
-const DynamicKakaoOAuthPage = dynamic(() => import('./page'), { ssr: false });
+import { useRouter, useSearchParams } from 'next/navigation';
+import { useEffect } from 'react';
 
-export default function KakaoOAuthPage() {
-  return <DynamicKakaoOAuthPage />;
-}
+const Page = () => {
+  const router = useRouter();
+  const searchParams = useSearchParams();
+
+  useEffect(() => {
+    const KAKAO_CODE = searchParams.get('code');
+
+    const postCode = async () => {
+      if (!KAKAO_CODE) {
+        console.error('코드가 유효하지 않습니다.');
+        return;
+      }
+      try {
+        const response = await fetch(`/api/kakao/oauth?code=${KAKAO_CODE}`, {
+          method: 'POST',
+          headers: {
+            'Content-Type': 'application/json',
+          },
+          body: JSON.stringify({ code: KAKAO_CODE }),
+        });
+
+        if (response.status === 200) {
+          router.push('/');
+        }
+      } catch (error) {
+        console.error('Error:', error);
+      }
+    };
+
+    postCode().catch((error) => {
+      console.error('Error:', error);
+    });
+  }, [router, searchParams]);
+
+  return <div>카카오톡으로 로그인중입니다.</div>;
+};
+
+export default Page;
